ml/requirements.txt
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
pyyaml==6.0.2
transformers==4.44.0
accelerate==0.34.0
ruff==0.6.9
